LPD-25380 merge tests from globalJS.spec.ts into js.spec.ts
